light/requirements.txt
diffusers==0.27.2
transformers==4.36.2
opencv-python
safetensors
pillow==10.2.0
einops
torch
peft
gradio==3.41.2
protobuf==3.20
openai>=1.0.0
basicsr>=1.4.2
realesrgan>=0.3.0
facexlib>=0.3.0
gfpgan>=1.3.8
